day 11 part 1 complete, enum.tostring generates the name of the field, not the value

+++ Advent Of Code 2019/Day11/PaintTracker.cs
@@ -109,7 +109,7 @@ namespace Day11
             // report the current colour to compiler
             if (visited.ContainsKey(current))
             {
-                Output.Add(visited[current].ToString());
+                Output.Add(((int)visited[current]).ToString());
             }
             else
             {
